Remove silly difforig command
I don't know why I had this. It's never useful because it only shows a
diff of the unsaved buffer versus the saved one, and by that point you
can either undo and if not you best be using a VCS. Which vim also has
commands for (i.e. DiffGitCached)